+++ src/Algos/Mads/OignonPollMethod.hpp
@@ -52,7 +52,7 @@ private:
 
     size_t n;// dimension
 
-	size_t nbOfPreviousFailure; // is the number of failure before this poll, used to derermine how many layers are created
+    int nbOfPreviousFailure; // is the number of failure before this poll, used to derermine how many layers are created
     
 };
